+++ ci-droid-tasks-consumer-services/src/test/java/com/societegenerale/cidroid/tasks/consumer/services/actionHandlers/PullRequestCleaningHandlerTest.java
@@ -1,18 +1,25 @@
 package com.societegenerale.cidroid.tasks.consumer.services.actionHandlers;
 
+import ch.qos.logback.classic.Logger;
+import ch.qos.logback.classic.LoggerContext;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.societegenerale.cidroid.tasks.consumer.services.RemoteGitHub;
 import com.societegenerale.cidroid.tasks.consumer.services.model.DateProvider;
 import com.societegenerale.cidroid.tasks.consumer.services.model.github.PullRequest;
 import com.societegenerale.cidroid.tasks.consumer.services.model.github.PushEvent;
+import com.societegenerale.cidroid.tasks.consumer.services.monitoring.TestAppender;
 import org.junit.Before;
 import org.junit.Test;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.Collections;
 
 import static com.societegenerale.cidroid.tasks.consumer.services.TestUtils.readFromInputStream;
+import static com.societegenerale.cidroid.tasks.consumer.services.monitoring.MonitoringEvents.OLD_PR_CLOSED;
+import static java.util.stream.Collectors.toList;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
 
 public class PullRequestCleaningHandlerTest {
@@ -25,7 +32,7 @@ public class PullRequestCleaningHandlerTest {
     private DateProvider dateProvider;
     private PushEvent pushEvent;
 
-
+    private TestAppender testAppender=new TestAppender();
 
     @Before
     public void setUp() throws IOException {
@@ -41,6 +48,10 @@ public class PullRequestCleaningHandlerTest {
 
         String pushEventPayload = readFromInputStream(getClass().getResourceAsStream("/pushEvent.json"));
         pushEvent = new ObjectMapper().readValue(pushEventPayload, PushEvent.class);
+
+        LoggerContext logCtx = (LoggerContext) LoggerFactory.getILoggerFactory();
+        Logger log = logCtx.getLogger("Main");
+        log.addAppender(testAppender);
     }
 
     @Test
@@ -54,6 +65,10 @@ public class PullRequestCleaningHandlerTest {
         String expectedRepositoryName = pushEvent.getRepository().getFullName();
         verify(remoteGitHub, times(1))
                 .closePullRequest(expectedRepositoryName, PULL_REQUEST_NUMBER);
+
+        assertThat(testAppender.events.stream()
+                .filter(logEvent -> logEvent.getMDCPropertyMap().getOrDefault("metricName", "NOT_FOUND").equals(OLD_PR_CLOSED)).findAny())
+                .isPresent();
     }
 
     @Test
@@ -66,6 +81,12 @@ public class PullRequestCleaningHandlerTest {
         pullRequestCleaningHandler.handle(pushEvent, Collections.singletonList(recentPullRequest));
 
         verifyZeroInteractions(remoteGitHub);
+
+        assertThat(testAppender.events.stream()
+                                      .filter(logEvent -> logEvent.getMDCPropertyMap().getOrDefault("metricName", "NOT_FOUND").equals(OLD_PR_CLOSED))
+                                      .collect(toList()))
+                .isEmpty();
+
     }
 
 }